requirements.txt
pylint
openai
python-dotenv
pydantic